requirements_local.txt
pyaes
pyasn1
python-socks
PyYAML
rsa
setuptools
Telethon
wheel
yt-dlp
f2
